Refactor components: move module.scss to styles directory
