+++ src/components/styles/InputCheckAll.module.scss
@@ -1,5 +1,5 @@
-@import '../../assets/styles/global/variables.scss';
-@import '../../assets/styles/global/mixin.scss';
+@import '../../../assets/styles/global/variables.scss';
+@import '../../../assets/styles/global/mixin.scss';
 
 .checkAll {
   position: absolute;
