+++ src/components/styles/TodosHeader.module.scss
@@ -1,5 +1,5 @@
-@import '../../assets/styles/global/variables.scss';
-@import '../../assets/styles/global/mixin.scss';
+@import '../../../assets/styles/global/variables.scss';
+@import '../../../assets/styles/global/mixin.scss';
 
 .header {
   position: absolute;
