+++ src/components/styles/TodosList.module.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/global/variables.scss';
+@import '../../../assets/styles/global/variables.scss';
 
 .list {
   width: 100%;
